+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Added checks and visual alerts on the login screen if the client has lost server and/or general Internet connectivity
+
 ## [2.2.10]
 ### Added
 - Added more thorough debugging / logging statements to almost all methods